requirements2.txt
arabic-reshaper==2.1.0
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-modeltranslation==0.15.2
django-resized==0.3.11
django-widget-tweaks==1.4.8
entrypoints==0.3
Flask==1.1.2
future==0.18.2
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.25
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.4
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.44
packaging==20.4
pandocfilters==1.4.3
parso==0.7.1
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
pycparser==2.20
Pygments==2.7.2
PyGoogleVoice==0.5.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-bidi==0.4.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
reportlab==3.5.53
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
scikit-learn==0.23.2
scipy==1.5.3
Send2Trash==1.5.0
six==1.15.0
sniffio==1.1.0
sqlparse==0.3.1
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==5.0.5
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
xhtml2pdf==0.2.5
xlwt==1.3.0
